pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pool-cli"
version = "0.1.0"
description = "Intelligent screenshot organizer — intention archaeology for your camera roll"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40.0",
    "google-generativeai>=0.8.0",
    "open-clip-torch>=2.26.0",
    "transformers>=4.30.0,<5.0.0",
    "torch>=2.1.0",
    "Pillow>=10.0.0",
    "pillow-heif>=0.16.0",
    "imagehash>=4.3.0",
    "numpy>=1.24.0",
    "umap-learn>=0.5.0",
    "hdbscan>=0.8.33",
    "ocrmac>=1.0.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
]

[project.scripts]
pool = "pool.cli:app"

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["pool*"]
